Manually enqueue view scripts fnd styles for any used blocks
Blocks rendered in the content are not actually having their scripts and
styles cached. This change caches the list of blocks used, and manually
enqueues assets for each of those blocks.

I don't think this should be necessary, but I couldn't find a
higher-level function which handles this in core.