Add hooks for extending the cached content functionality
Some plugins trigger actions on the content that rely on global state
outside of the blocks themselves. One example is advertising plugins
like Advanced Ads. This change adds a filter which can be used to extend
the data stored with a post, and an action which can perform additional
setup and output based on this cached data.